requirements.txt
tyro==0.8.14
wandb==0.18.5
jsonlines==4.0.0
transformers==4.46.1
scikit-learn==1.5.2
numpy==1.26.4
pandas
torch==2.5.1
peft==0.13.2
datasets==3.1.0
openai-1.53.1
